fix: correct sensor parsing and data fidelity issues
- Store humidity as Float with Room migration to v3 (keeps fractional %)
- Parse BTHome v2 extended type IDs 0x10/0x11 as 2-byte measurements
- Fix ATC 12-byte payloads: uint16 humidity at offsets 8-9, battery from voltage only
- Re-arm threshold alerts on a 1h cooldown while condition persists
- Use lastDbTimestamp when correcting downloaded history timestamps
- Null-check GATT notification descriptor before write
- Re-acquire WakeLock if it expires during long service runs
- Refresh dashboard chart endTime every minute so new data appears
- Use vector thermometer icon for notifications; drop hardcoded GERMANY locale
- Upgrade AGP to 8.9.2 + buildTools 36.0.0 (compileSdk 36 support)
- Add ACCESS_BACKGROUND_LOCATION and fix README engine/dependency wording

+++ app/src/main/java/com/example/readmymi/data/SensorEntity.kt
@@ -9,7 +9,7 @@ data class SensorEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val macAddress: String,
     val temperature: Float,
-    val humidity: Int,
+    val humidity: Float,
     val battery: Int,
     val timestamp: Long
 ) {
